refactoring: extract tree visualization method
